etl/requirements.txt
cassandra-driver==3.25.0
click==8.1.3
geomet==0.2.1.post1
numpy==1.23.3
pandas==1.5.0
pip==22.2.2
python-dateutil==2.8.2
pytz==2022.2.1
setuptools==65.2.0
six==1.16.0
wheel==0.37.1
